Rename to PatternViewModel, move poylines and stops to the patterns package
