+++ app/src/main/java/rectangledbmi/com/pittsburghrealtimetracker/patterns/stops/StopView.java
@@ -1,4 +1,4 @@
-package rectangledbmi.com.pittsburghrealtimetracker.stops;
+package rectangledbmi.com.pittsburghrealtimetracker.patterns.stops;
 
 import rx.Observer;
